+++ ntex/src/server/service.rs
@@ -85,7 +85,7 @@ where
         match req {
             ServerMessage::Connect(stream) => {
                 let stream = FromStream::from_stream(stream).map_err(|e| {
-                    error!("Can not convert to an async tcp stream: {}", e);
+                    error!("Can not convert to an async io stream: {}", e);
                 });
 
                 if let Ok(stream) = stream {